+++ ktrain/core.py
@@ -174,7 +174,14 @@ class Learner(ABC):
             val = val_data
         else:
             val = self.val_data
-        if U.is_multilabel(val):
+
+        classification, multilabel = U.is_classifier(self.model)
+        if not classification:
+            warnings.warn('learner.validate is only for classification problems. ' 
+                          'For regression, etc., use learner.predict and learner.ground_truth '
+                          'to manually validate.')
+            return
+        if U.is_multilabel(val) or multilabel:
             warnings.warn('multilabel confusion matrices not yet supported')
             return
         y_pred = self.predict(val_data=val)
@@ -192,12 +199,16 @@ class Learner(ABC):
         return cm
 
 
-    def top_losses(self, n=8, val_data=None):
+    def top_losses(self, n=8, val_data=None, preproc=None):
         """
         Computes losses on validation set sorted by examples with top losses
         Args:
           n(int or tuple): a range to select in form of int or tuple
                           e.g., n=8 is treated as n=(0,8)
+          preproc (Preprocessor): A TextPreprocessor or ImagePreprocessor.
+                                  For some data like text data, a preprocessor
+                                  is required to undo the pre-processing
+                                   to correctly view raw data.
           val_data:  optional val_data to use instead of self.val_data
         Returns:
             list of n tuples where first element is either 
@@ -206,6 +217,8 @@ class Learner(ABC):
 
         """
         import tensorflow as tf
+        #multilabel = True if U.is_multilabel(val) else False
+        classification, multilabel = U.is_classifier(self.model)
 
         # check validation data and arguments
         if val_data is not None:
@@ -216,24 +229,35 @@ class Learner(ABC):
         if type(n) == type(42):
             n = (0, n)
 
-        multilabel = True if U.is_multilabel(val) else False
 
         # get predicictions and ground truth
         y_pred = self.predict(val_data=val)
         y_true = self.ground_truth(val_data=val)
         y_true = y_true.astype('float32')
 
+
         # compute loss
         losses = self.model.loss_functions[0](tf.convert_to_tensor(y_true), tf.convert_to_tensor(y_pred))
         losses = tf.Session().run(losses)
+        class_names = [] if preproc is None else preproc.get_classes()
+        if preproc is None: 
+            class_fcn = lambda x:x
+        else:
+            class_fcn = lambda x:class_names[x]
+
+        # don't put regression predictions in a list
+        if not classification and len(y_pred.shape) == 2 and y_pred.shape[1] == 1:
+            y_pred = np.squeeze(y_pred)
+            y_pred = np.around(y_pred, 2)
 
         # sort by loss and prune correct classifications, if necessary
-        if 'acc' in self.model.metrics_names and not multilabel:
+        if classification and not multilabel:
             y_p = np.argmax(y_pred, axis=1)
             y_t = np.argmax(y_true, axis=1)
-            tups = [(i,x) for i, x in enumerate(losses) if y_p[i] != y_t[i]]
+            tups = [(i,x, class_fcn(y_t[i]), class_fcn(y_p[i])) for i, x in enumerate(losses) 
+                     if y_p[i] != y_t[i]]
         else:
-            tups = [(i,x) for i, x in enumerate(losses)]
+            tups = [(i,x, y_true[i], np.around(y_pred[i],2)) for i, x in enumerate(losses)]
         tups.sort(key=operator.itemgetter(1), reverse=True)
 
         # prune by given range
@@ -241,16 +265,16 @@ class Learner(ABC):
         return tups
 
 
-    def view_top_losses(self, preproc=None, n=8, val_data=None):
+    def view_top_losses(self, n=8, preproc=None, val_data=None):
         """
         Views observations with top losses in validation set.
         Args:
+         n(int or tuple): a range to select in form of int or tuple
+                          e.g., n=8 is treated as n=(0,8)
          preproc (Preprocessor): A TextPreprocessor or ImagePreprocessor.
                                  For some data like text data, a preprocessor
                                  is required to undo the pre-processing
                                  to correctly view raw data.
-         n(int or tuple): a range to select in form of int or tuple
-                          e.g., n=8 is treated as n=(0,8)
           val_data:  optional val_data to use instead of self.val_data
         Returns:
             list of n tuples where first element is either 
@@ -258,6 +282,8 @@ class Learner(ABC):
             is loss.
 
         """
+        # TODO: fix this mess
+
         # check validation data and arguments
         if val_data is not None:
             val = val_data
@@ -265,33 +291,47 @@ class Learner(ABC):
             val = self.val_data
         if val is None: raise Exception('val_data must be supplied to get_learner or view_top_losses')
 
-        # TODO: fix this
-        tups = self.top_losses(n=n, val_data=val)
+        # get top losses and associated data
+        tups = self.top_losses(n=n, val_data=val, preproc=preproc)
+
+        # get multilabel status and class names
+        classes = preproc.get_classes() if preproc is not None else None
+
+        # iterate through losses
         for tup in tups:
+
+            # get data
             idx = tup[0]
             loss = tup[1]
+            truth = tup[2]
+            pred = tup[3]
+
+            # Image Classification
             if type(val).__name__ in ['DirectoryIterator', 'DataFrameIterator']:
-                # idx is replaced with a file path
-                # replace IDs with file paths, if possible
                 fpath = val.filepaths[tup[0]]
+                fp = os.path.join(os.path.basename(os.path.dirname(fpath)), os.path.basename(fpath))
                 plt.figure()
-                plt.title("%s (LOSS: %s)" % (fpath, round(loss,3)))
+                plt.title("%s | loss:%s | true:%s | pred:%s)" % (fp, round(loss,2), truth, pred))
                 show_image(fpath)
             else:
+                # Image Classification from Array
                 if type(val).__name__ in ['NumpyArrayIterator']:
                     obs = val.x[idx]
                     if preproc is not None: obs = preproc.undo(obs)
                     plt.figure()
-                    plt.title("id:%s (LOSS: %s)" % (idx, round(loss,3)))
+                    plt.title("id:%s | loss:%s | true:%s | pred:%s)" % (idx, round(loss,2), truth, pred))
                     plt.imshow(np.squeeze(obs))
+                # everything else including text classification
                 else:
                     if U.bert_data_tuple(val):
                         obs = val[0][0][idx]
                     else:
                         obs = val[0][idx]
                     if preproc is not None: obs = preproc.undo(obs)
+                    if type(obs) == str:
+                        obs = ' '.join(obs.split()[:512])
                     print('----------')
-                    print("val_id:%s (LOSS: %s)\n" % (idx, round(loss,3)))
+                    print("id:%s | loss:%s | true:%s | pred:%s)\n" % (idx, round(loss,2), truth, pred))
                     print(obs)
         return