+++ ktrain/imports.py
@@ -34,7 +34,8 @@ import requests
 # verify=False added to avoid headaches from some corporate networks
 from requests.packages.urllib3.exceptions import InsecureRequestWarning
 requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
-
+import eli5
+from eli5.lime import TextExplainer
 
 
 
@@ -69,6 +70,7 @@ from keras.preprocessing.image import ImageDataGenerator
 from keras.preprocessing.text import Tokenizer
 from keras.utils import Sequence, to_categorical
 from keras.utils import multi_gpu_model
+from keras.activations import sigmoid
 
 from keras.applications.resnet50 import ResNet50
 from keras.applications.mobilenet import MobileNet