+++ ktrain/tests/test_imageclassification.py
@@ -52,6 +52,7 @@ class TestImageClassification(TestCase):
 
         # test top_losses
         obs = learner.top_losses(n=1, val_data=val)
+        print(obs)
         if obs:
             self.assertIn(obs[0][0], list(range(U.nsamples_from_data(val))))
         else:
@@ -76,9 +77,15 @@ class TestImageClassification(TestCase):
         # test predictor
         p = ktrain.get_predictor(learner.model, preproc)
         r = p.predict_folder('image_data/image_folder/train/')
+        print(r)
         self.assertEqual(r[0][1], 'cat')
+        r = p.predict_proba_folder('image_data/image_folder/train/')
+        self.assertEqual(np.argmax(r[0][1]), 0)
         r = p.predict_filename('image_data/image_folder/train/cat/cat.11737.jpg')
         self.assertEqual(r, ['cat'])
+        r = p.predict_proba_filename('image_data/image_folder/train/cat/cat.11737.jpg')
+        self.assertEqual(np.argmax(r), 0)
+
         p.save('/tmp/test_predictor')
         p = ktrain.load_predictor('/tmp/test_predictor')
         r = p.predict_filename('image_data/image_folder/train/cat/cat.11737.jpg')