+++ ktrain/tests/test_textclassification.py
@@ -7,11 +7,12 @@ import os
 os.environ["CUDA_DEVICE_ORDER"]="PCI_BUS_ID";
 os.environ["CUDA_VISIBLE_DEVICES"]="0"
 sys.path.insert(0,'../..')
+import IPython
 from unittest import TestCase, main, skip
 import numpy as np
 import ktrain
 from ktrain import text as txt
-
+TEST_DOC = 'god christ jesus mother mary church sunday lord heaven amen'
 
 class TestTextClassification(TestCase):
 
@@ -79,13 +80,12 @@ class TestTextClassification(TestCase):
 
         # test predictor
         p = ktrain.get_predictor(learner.model, preproc)
-        self.assertEqual(p.predict(['god christ jesus mother mary church sunday lord heaven amen'])[0], 
-                         'soc.religion.christian')
+        self.assertEqual(p.predict([TEST_DOC])[0], 'soc.religion.christian')
         p.save('/tmp/test_predictor')
         p = ktrain.load_predictor('/tmp/test_predictor')
-        self.assertEqual(p.predict(['god christ jesus mother mary church sunday lord heaven amen'])[0], 
-                         'soc.religion.christian')
-
+        self.assertEqual(p.predict(TEST_DOC), 'soc.religion.christian')
+        self.assertEqual(np.argmax(p.predict_proba([TEST_DOC])[0]), 3)
+        self.assertEqual(type(p.explain(TEST_DOC)), IPython.core.display.HTML)
 
 
     def test_nbsvm(self):
@@ -132,13 +132,12 @@ class TestTextClassification(TestCase):
 
         # test predictor
         p = ktrain.get_predictor(learner.model, preproc)
-        self.assertEqual(p.predict(['god christ jesus mother mary church sunday lord heaven amen'])[0], 
-                         'soc.religion.christian')
+        self.assertEqual(p.predict([TEST_DOC])[0], 'soc.religion.christian')
         p.save('/tmp/test_predictor')
         p = ktrain.load_predictor('/tmp/test_predictor')
-        self.assertEqual(p.predict(['god christ jesus mother mary church sunday lord heaven amen'])[0], 
-                         'soc.religion.christian')
-
+        self.assertEqual(p.predict(TEST_DOC), 'soc.religion.christian')
+        self.assertEqual(np.argmax(p.predict_proba([TEST_DOC])[0]), 3)
+        self.assertEqual(type(p.explain(TEST_DOC)), IPython.core.display.HTML)
 
     def test_logreg(self):
         trn, val, preproc = txt.texts_from_array(x_train=self.trn[0], 
@@ -184,14 +183,15 @@ class TestTextClassification(TestCase):
 
         # test predictor
         p = ktrain.get_predictor(learner.model, preproc)
-        self.assertEqual(p.predict(['god christ jesus mother mary church sunday lord heaven amen'])[0], 
-                         'soc.religion.christian')
+        self.assertEqual(p.predict([TEST_DOC])[0], 'soc.religion.christian')
         p.save('/tmp/test_predictor')
         p = ktrain.load_predictor('/tmp/test_predictor')
-        self.assertEqual(p.predict(['god christ jesus mother mary church sunday lord heaven amen'])[0], 
-                         'soc.religion.christian')
+        self.assertEqual(p.predict(TEST_DOC), 'soc.religion.christian')
+        self.assertEqual(np.argmax(p.predict_proba([TEST_DOC])[0]), 3)
+        self.assertEqual(type(p.explain(TEST_DOC)), IPython.core.display.HTML)
 
 
+    #@skip('temporarily disabled')
     def test_bigru(self):
         trn, val, preproc = txt.texts_from_array(x_train=self.trn[0], 
                                                  y_train=self.trn[1],
@@ -209,7 +209,7 @@ class TestTextClassification(TestCase):
 
         # test training results
         self.assertAlmostEqual(max(hist.history['lr']), lr)
-        self.assertGreater(max(hist.history['val_acc']), 0.92)
+        self.assertGreater(max(hist.history['val_acc']), 0.9)
         self.assertAlmostEqual(max(hist.history['momentum']), 0.95)
         self.assertAlmostEqual(min(hist.history['momentum']), 0.85)
 
@@ -238,14 +238,16 @@ class TestTextClassification(TestCase):
 
         # test predictor
         p = ktrain.get_predictor(learner.model, preproc)
-        self.assertEqual(p.predict(['god christ jesus mother mary church sunday lord heaven amen'])[0], 
-                         'soc.religion.christian')
+        self.assertEqual(p.predict([TEST_DOC])[0], 'soc.religion.christian')
         p.save('/tmp/test_predictor')
         p = ktrain.load_predictor('/tmp/test_predictor')
-        self.assertEqual(p.predict(['god christ jesus mother mary church sunday lord heaven amen'])[0], 
-                         'soc.religion.christian')
+        self.assertEqual(p.predict([TEST_DOC])[0], 'soc.religion.christian')
+        self.assertEqual(p.predict(TEST_DOC), 'soc.religion.christian')
+        self.assertEqual(np.argmax(p.predict_proba([TEST_DOC])[0]), 3)
+        self.assertEqual(type(p.explain(TEST_DOC)), IPython.core.display.HTML)
 
 
+    #@skip('temporarily disabled')
     def test_bert(self):
         trn, val, preproc = txt.texts_from_array(x_train=self.trn[0], 
                                                  y_train=self.trn[1],
@@ -262,7 +264,7 @@ class TestTextClassification(TestCase):
 
         # test training results
         self.assertAlmostEqual(max(hist.history['lr']), lr)
-        self.assertGreater(max(hist.history['acc']), 0.75)
+        self.assertGreater(max(hist.history['acc']), 0.7)
 
 
         # test top losses
@@ -288,12 +290,14 @@ class TestTextClassification(TestCase):
 
         # test predictor
         p = ktrain.get_predictor(learner.model, preproc)
-        self.assertEqual(p.predict(['god christ jesus mother mary church sunday lord heaven amen'])[0], 
-                      'soc.religion.christian')
+        self.assertEqual(p.predict([TEST_DOC])[0], 'soc.religion.christian')
         p.save('/tmp/test_predictor')
         p = ktrain.load_predictor('/tmp/test_predictor')
-        self.assertEqual(p.predict(['god christ jesus mother mary church sunday lord heaven amen'])[0], 
-                         'soc.religion.christian')
+        self.assertEqual(p.predict(TEST_DOC), 'soc.religion.christian')
+        self.assertEqual(np.argmax(p.predict_proba([TEST_DOC])[0]), 3)
+        self.assertEqual(type(p.explain(TEST_DOC)), IPython.core.display.HTML)
+
+
 
 
 if __name__ == "__main__":